refactor: update imports fo types from a centralized place and clean up duplicated type definitions for modularity and reusability
